Fixed 64 bit version bug uncovered by pm_kan with gcc 4.8
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@5267 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/core/ItemData.cpp
@@ -1774,7 +1774,7 @@ bool CItemData::DeSerializePlainText(const std::vector<char> &v)
 
   while (iter != v.end()) {
     int type = (*iter++ & 0xff); // required since enum is an int
-    if (distance(v.end(), iter) < sizeof(size_t)) {
+    if (distance(v.end(), iter) < sizeof(uint32)) {
       ASSERT(0); // type must ALWAYS be followed by length
       return false;
     }
